Enables hostmonitor to work on pacemaker-remote.
Masakari-hostmonitor can check other host's status with pacemaker-remote.
Previous hostmonitor only check a host which runs on the full cluster stack
of corosync. This hits the upper limit of the cluster, 16 hosts, and we
have to set up multi pacemaker cluster when the number of host is >16.
Pacemaker-remote resolves the problem, so to enable hostmonitor to check
a cluster status by pacemaker-remote we can set up a number of hosts in
one cluster more than 16.